refactor: streamline entry point and package node retrieval in ng-package
- Introduced helper functions `findEntryPointInProgress` and `findPackageNode` to simplify the retrieval of entry point and package nodes in the build graph.
- Updated various transforms to utilize these new functions, improving code readability and maintainability.
- Removed redundant error handling for node retrieval, centralizing it within the new helper functions.
- Enhanced bundler warning handling in rollup integration for better logging of non-critical warnings.
