+++ src/lib/ng-package/entry-point/entry-point.transform.ts
@@ -1,9 +1,8 @@
 import { pipe, tap } from 'rxjs';
-import { ComplexPredicate } from '../../graph/build-graph';
-import { Node, STATE_DONE } from '../../graph/node';
+import { STATE_DONE } from '../../graph/node';
 import { Transform } from '../../graph/transform';
 import * as log from '../../utils/log';
-import { EntryPointNode, isEntryPointInProgress } from '../nodes';
+import { findEntryPointInProgress } from '../nodes';
 
 /**
  * A re-write of the `transformSources()` script that transforms an entry point from sources to distributable format.
@@ -34,10 +33,7 @@ export const entryPointTransformFactory = (compileTs: Transform, writeBundles: T
   pipe(
     tap(graph => {
       // Peek the first entry point from the graph
-      const entryPoint = graph.find(isEntryPointInProgress() as ComplexPredicate<Node, EntryPointNode>);
-      if (!entryPoint) {
-        throw new Error('Could not find entry point in progress');
-      }
+      const entryPoint = findEntryPointInProgress(graph);
       log.msg('\n------------------------------------------------------------------------------');
       log.msg(`Building entry point '${entryPoint.data.entryPoint.moduleId}'`);
       log.msg('------------------------------------------------------------------------------');
@@ -48,10 +44,7 @@ export const entryPointTransformFactory = (compileTs: Transform, writeBundles: T
     writeBundles,
     writePackage,
     tap(graph => {
-      const entryPoint = graph.find(isEntryPointInProgress() as ComplexPredicate<Node, EntryPointNode>);
-      if (!entryPoint) {
-        throw new Error('Could not find entry point in progress');
-      }
+      const entryPoint = findEntryPointInProgress(graph);
       entryPoint.state = STATE_DONE;
     }),
   );
