+++ src/lib/ng-package/entry-point/write-package.transform.ts
@@ -10,7 +10,7 @@ import { colors } from '../../utils/color';
 import { copyFile, mkdir, rmdir, stat, writeFile } from '../../utils/fs';
 import * as log from '../../utils/log';
 import { ensureUnixPath } from '../../utils/path';
-import { EntryPointNode, PackageNode, fileUrl, isEntryPoint, isPackage } from '../nodes';
+import { EntryPointNode, PackageNode, fileUrl, findPackageNode, isEntryPoint } from '../nodes';
 import { NgPackagrOptions } from '../options.di';
 import { NgPackage } from '../package';
 import { NgEntryPoint } from './entry-point';
@@ -26,10 +26,7 @@ export const writePackageTransform = (options: NgPackagrOptions): Transform =>
       throw new Error('Could not find entry point in progress');
     }
     const ngEntryPoint: NgEntryPoint = entryPoint.data.entryPoint;
-    const ngPackageNode: PackageNode | undefined = graph.find(isPackage);
-    if (!ngPackageNode) {
-      throw new Error('Could not find package node in build graph');
-    }
+    const ngPackageNode = findPackageNode(graph);
     const ngPackage = ngPackageNode.data;
     const { destinationFiles } = entryPoint.data;
