+++ src/lib/ng-package/nodes.ts
@@ -1,7 +1,7 @@
 import type { NgtscProgram, ParsedConfiguration, Program } from '@angular/compiler-cli';
 import ts from 'typescript';
 import { FileCache } from '../file-system/file-cache';
-import { ComplexPredicate } from '../graph/build-graph';
+import { BuildGraph, ComplexPredicate } from '../graph/build-graph';
 import { Node } from '../graph/node';
 import { by, isInProgress, isPending } from '../graph/select';
 import { AngularDiagnosticsCache } from '../ngc/angular-diagnostics-cache';
@@ -36,6 +36,24 @@ export function isEntryPointInProgress(): ComplexPredicate<Node, EntryPointNode>
 
 export function isEntryPointPending(): ComplexPredicate<Node, EntryPointNode> {
   return by(n => isEntryPoint(n) && isPending(n)) as ComplexPredicate<Node, EntryPointNode>;
+}
+
+export function findEntryPointInProgress(graph: BuildGraph): EntryPointNode {
+  const entryPoint = graph.find(isEntryPointInProgress() as ComplexPredicate<Node, EntryPointNode>);
+  if (!entryPoint) {
+    throw new Error('Could not find entry point in progress');
+  }
+
+  return entryPoint;
+}
+
+export function findPackageNode(graph: BuildGraph): PackageNode {
+  const packageNode: PackageNode | undefined = graph.find(isPackage);
+  if (!packageNode) {
+    throw new Error('Could not find package node in build graph');
+  }
+
+  return packageNode;
 }
 
 export function fileUrl(path: string): string {
