+++ src/lib/ngc/compile-source-files.ts
@@ -1,9 +1,8 @@
 import { CompilerOptions, ParsedConfiguration } from '@angular/compiler-cli';
 import { join } from 'node:path';
 import ts from 'typescript';
-import { BuildGraph, ComplexPredicate } from '../graph/build-graph';
-import { Node } from '../graph/node';
-import { EntryPointNode, PackageNode, isEntryPointInProgress, isPackage } from '../ng-package/nodes';
+import { BuildGraph } from '../graph/build-graph';
+import { findEntryPointInProgress, findPackageNode } from '../ng-package/nodes';
 import { NgPackagrOptions } from '../ng-package/options.di';
 import { StylesheetProcessor } from '../styles/stylesheet-processor';
 import { augmentProgramWithVersioning, cacheCompilerHost } from '../ts/cache-compiler-host';
@@ -20,14 +19,8 @@ export async function compileSourceFiles(
   const { NgtscProgram, formatDiagnostics } = await import('@angular/compiler-cli');
   const { cacheDirectory, watch, cacheEnabled } = options;
   const tsConfigOptions: CompilerOptions = { ...tsConfig.options, ...extraOptions };
-  const entryPoint: EntryPointNode | undefined = graph.find(isEntryPointInProgress() as ComplexPredicate<Node, EntryPointNode>);
-  if (!entryPoint) {
-    throw new Error('Could not find entry point in progress');
-  }
-  const ngPackageNode: PackageNode | undefined = graph.find(isPackage);
-  if (!ngPackageNode) {
-    throw new Error('Could not find package node in build graph');
-  }
+  const entryPoint = findEntryPointInProgress(graph);
+  const ngPackageNode = findPackageNode(graph);
   const inlineStyleLanguage = ngPackageNode.data.inlineStyleLanguage;
 
   const cacheDir = cacheEnabled && cacheDirectory;
